Add subscription protection for server routes

+++ fitness-app/src/server/exercises/deleteExercise.ts
@@ -2,10 +2,17 @@
 
 import { prisma } from "@/lib/prisma";
 import { getOrCreateCurrentUser } from "../users/getOrCreateCurrentUser";
+import { getUserSubscription } from "../subscription/getUserSubscription";
+import { redirect } from "next/navigation";
 
 export async function deleteExercise(exerciseId : string) {
 
     await getOrCreateCurrentUser();
+    const subscription = await getUserSubscription();
+
+    if (!subscription.subscriptionActive) {
+      redirect("/pricing");
+    }
 
     await prisma.set.deleteMany({
         where: {
